fix CI: drain BackendRig's Qt event loop before executor teardown; drop useless cast
Kanban/ThreadSanitizer, Linux/clang-coverage, and Linux/all-optional-
features(clang) all failed on the same underlying bug: a heap-use-
after-free in QtExecutor::post()/CompletionState::setException,
confirmed by TSan and reproduced as a plain SIGSEGV in the two
uninstrumented builds.

Root cause: QtExecutor::post() (qt_executor.hpp) uses
Qt::QueuedConnection -- it enqueues a callback on the Qt event loop and
returns immediately, it does not run it. ThreadPoolExecutor's
destructor joins every worker thread, which guarantees every post()
call was *made* before it returns, but not that the resulting event
was *pumped*. Bridge::executeVia (bridge.hpp) chains two Completion
objects per action, so a single failed action can enqueue a *second*,
nested post() from inside the first posted callback -- invisible to a
caller's own pumpUntil(outstanding == 0), which only tracks the outer
completion. BackendRig could observe done and tear itself down,
freeing _qtExecutor, while that inner post's callback was still
sitting undelivered in the Qt event queue; when it finally ran, it
touched freed memory.

Fix: BackendRig::~BackendRig() now resets _workerPool explicitly
(forcing every pool-issued post() to have already happened) and then
drains the Qt event loop for a few bounded slices -- long enough for a
nested post to both arrive and run -- before the rest of member
destruction (including _qtExecutor) proceeds. This mirrors
QtWebSocketServer::closeGracefully()'s own established
processEvents-drain pattern, just applied to the worker-pool/executor
pair instead of the socket server.

Also drops a redundant static_cast<int> in
test_board_concurrent_drag.cpp that GCC's -Wuseless-cast (clang has no
equivalent warning) correctly flagged as dead code: the expression was
already int-typed before the outer cast.

+++ examples/kanban/tests/test_board_concurrent_drag.cpp
@@ -254,7 +254,7 @@ TEST_CASE("Concurrent BoardBridge::moveTask calls (N=4) never desync positions",
 
         for (int a = 0; a < kMovesPerBridge; ++a) {
             const QString destColumn = columns[static_cast<std::size_t>(a) % columns.size()];
-            const auto position = static_cast<int>((a * 3 + static_cast<int>(i)) % static_cast<int>(kClients));
+            const auto position = (a * 3 + static_cast<int>(i)) % static_cast<int>(kClients);
             ++outstanding;
             bridge.moveTask(taskIds[i], destColumn, swimlaneId, position);
         }
